helloworld_project/requirements.txt
asgiref==3.7.2
blessed==1.20.0
botocore==1.31.62
cement==2.8.2
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.0
colorama==0.4.3
distlib==0.3.7
Django==4.2.5
djangorestframework==3.14.0
filelock==3.12.4
identify==2.5.29
idna==3.4
jmespath==1.0.1
nodeenv==1.8.0
pathspec==0.10.1
platformdirs==3.10.0
pre-commit==3.4.0
psycopg2==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.4
termcolor==1.1.0
typing_extensions==4.8.0
urllib3==1.26.17
virtualenv==20.24.5
wcwidth==0.1.9
